+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 ## 1.30.0 [unreleased]
 
+### Features
+1. [#440](https://github.com/influxdata/influxdb-client-python/pull/440): Add possibility to specify timestamp column and its timezone [DataFrame]
+
 ## 1.29.1 [2022-05-23]
 
 ### Bug Fixes